+++ BranchUnityTestBed/Assets/Branch/Branch.cs
@@ -8,7 +8,7 @@ using System.Runtime.InteropServices;
 public class Branch : MonoBehaviour
 {
 
-	public static string sdkVersion = "2.0.0";
+	public static string sdkVersion = "2.1.0";
 
 	public delegate void BranchCallbackWithParams(Dictionary<string, object> parameters, string error);
 	public delegate void BranchCallbackWithUrl(string url, string error);
@@ -256,6 +256,7 @@ public class Branch : MonoBehaviour
 		_clearPartnerParameters();
 	}
 
+	[Obsolete("setTrackingDisabled() is deprecated. Please use setConsumerProtectionAttributionLevel() instead.")]
 	public static void setTrackingDisabled(bool value)
 	{
 		_setTrackingDisabled(value);
@@ -330,6 +331,34 @@ public class Branch : MonoBehaviour
 	public static void validate()
 	{
 		_validate();
+	}
+
+	#endregion
+
+	#region Google On Device Measurement methods
+
+	public static void setSDKWaitTimeForThirdPartyAPIs(double waitTime)
+	{
+		_setSDKWaitTimeForThirdPartyAPIs(waitTime);
+	}
+
+	public static void setODMInfo(string odmInfo, double firstOpenTimestamp)
+	{
+		_setODMInfo(odmInfo, firstOpenTimestamp);
+	}
+
+	public static void setAnonID(string anonID)
+	{
+		_setAnonID(anonID);
+	}
+
+	#endregion
+	
+	#region Set Consumer Protection Attribution methods
+
+	public static void setConsumerProtectionAttributionLevel(string level)
+	{
+		_setConsumerProtectionAttributionLevel(level);
 	}
 
 	#endregion
@@ -475,6 +504,18 @@ public class Branch : MonoBehaviour
 
 	[DllImport ("__Internal")]
 	private static extern void _validate();
+
+	[DllImport ("__Internal")]
+	private static extern void _setSDKWaitTimeForThirdPartyAPIs(double waitTime);
+
+	[DllImport ("__Internal")]
+	private static extern void _setODMInfo(string odmInfo, double firstOpenTimestamp);
+
+	[DllImport ("__Internal")]
+	private static extern void _setAnonID(string anonID);
+	
+	[DllImport ("__Internal")]
+	private static extern void _setConsumerProtectionAttributionLevel(string level);
 	    
 #elif UNITY_ANDROID && !UNITY_EDITOR
 
@@ -590,6 +631,22 @@ public class Branch : MonoBehaviour
 		BranchAndroidWrapper.validate();
 	}
 
+	private static void _setSDKWaitTimeForThirdPartyAPIs(double waitTime) {
+		BranchAndroidWrapper.setSDKWaitTimeForThirdPartyAPIs(waitTime);
+	}
+
+	private static void _setODMInfo(string odmInfo, double firstOpenTimestamp) {
+		BranchAndroidWrapper.setODMInfo(odmInfo, firstOpenTimestamp);
+	}
+
+	private static void _setAnonID(string anonID) {
+		BranchAndroidWrapper.setAnonID(anonID);
+	}
+
+	private static void _setConsumerProtectionAttributionLevel(string level) {
+		BranchAndroidWrapper.setConsumerProtectionAttributionLevel(level);
+	}
+
 #else
 
 	private static void _setFBAppID(string facebookAppID) { }
@@ -677,6 +734,14 @@ public class Branch : MonoBehaviour
 	}
 
 	private static void _validate() { }
+
+	private static void _setSDKWaitTimeForThirdPartyAPIs(double waitTime) {}
+
+	private static void _setODMInfo(string odmInfo, double firstOpenTimestamp) {}
+
+	private static void _setAnonID(string anonID) {}
+	
+	private static void _setConsumerProtectionAttributionLevel(string level) {}
 
 	private static void callNotImplementedCallbackForParamCallback(string callbackId)
 	{